Fixed sonar

+++ src/main/java/g3401_3500/s3449_maximize_the_minimum_game_score/Solution.java
@@ -23,7 +23,7 @@ public class Solution {
                     tr = 0;
                     skip++;
                 } else {
-                    long cur = tr * (long) p;
+                    long cur = tr * p;
                     long ops = (mid - cur + p - 1L) / p;
                     tot += 2 * ops - 1 + skip;
                     tr = Math.max(ops - 1, 0);
